refactor: extract PermissionAndPreflightTypes.swift from ServiceProtocols.swift
Extracts 9 permission/preflight-related types (5 enums + 2 structs + more)
from ServiceProtocols.swift into a dedicated PermissionAndPreflightTypes.swift:
- enum MicrophonePermissionState (3-9)
- enum MicrophonePermissionStatus (10-19)
- struct MicrophoneRecoveryGuidance (20-25)
- enum RecordingStartPreflightResult (26-41)
- enum ScreenCapturePermissionState (42-48)
- enum ScreenCapturePermissionStatus (49-57)
- struct ScreenCaptureRecoveryGuidance (58-62)
- enum ScreenshotCapturePreflightResult (63-78)
- enum ScreenshotSelectionResult (79-83)

Byte-identical move; SHA-256 3eac6cd22257bd74494889dc54b737deb7095cb827b58261a3ed676daeec0fa1
verified against origin/main:ServiceProtocols.swift lines 3-83.

Reduces ServiceProtocols.swift from 290 to 210 lines. Only the 20+ protocol
definitions remain in ServiceProtocols.swift — cleaner separation of "types
services use" from "protocols services conform to".

Tests: 667 (unchanged).
